Extract process runs helper methods
